backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
langchain
langchain-openai
langgraph
python-dotenv
requests
reportlab
